chore(starknet-sequencer-adapter): update sequencer (#568)
chore(straknet-sequencer-adapter): update sequencer

+++ generate_sqlx.sh
@@ -1,21 +1,39 @@
 #!/bin/bash
 
-folders=("ark-marketplace-api" "ark-marketplace-cron" "ark-orderbook-api")
+# Dont start the script if DBS are not set 
+if [[ -n "${MARKETPLACE_DATABASE_URL}"  || -n "${ORDERBOOK_DATABASE_URL}" ]]; then
+  marketplace_db=$MARKETPLACE_DATABASE_URL
+  orderbook_db=$ORDERBOOK_DATABASE_URL
+  folders=("ark-marketplace-api" "ark-marketplace-cron" "ark-orderbook-api" "ark-indexer-transactions")
+  rm -rf .sqlx/*
+  for folder in "${folders[@]}"
+  do
+    default_db=$marketplace_db
+    if [ "$folder" = "ark-orderbook-api" ]; then 
+      default_db=$orderbook_db
+      echo "run on orderbook DB" 
+    else
+        echo "run on marketplace DB" 
+    fi;
+    cd $folder
 
-rm -rf .sqlx/*
+    cargo sqlx prepare --database-url $default_db
 
-for folder in "${folders[@]}"
-do
-  cd $folder
+    cd ..
 
-  cargo sqlx prepare
+    cp -R $folder/.sqlx/* .sqlx/
+  done
 
-  cd ..
+  cd .sqlx
+  # git add .
+  # git commit -am "feat(sqlx): add sqlx files"
+  echo "check files before commit"
+else
+  echo "You should provide MARKETPLACE_DATABASE_URL && ORDERBOOK_DATABASE_URL environement variable at least"
+fi
 
-  cp -R $folder/.sqlx/* .sqlx/
-done
 
-cd .sqlx
-git add .
-git commit -am "feat(sqlx): add sqlx files"
-echo "Files committed"
+
+
+
+
